Implemented RegisterCredentials
The registerCredentials method required the adjustment of the SuccessCode class to include a method for finding a specific successcode associated with a given numerical value for the purposes of determining the response. All of the values in the enum were changed to decimal values to reflect the responses received from the server and the method was made static to allow for static reference and help avoid unnecessary dependencies. No tests currently exist for this, they'll be implemented in the next commit.