Adjusted ViewModel-ServerCommunicator Relationship
In order to facilitate proper communication between the client and the server, I added an overloaded constructor to the view models that allows us to differentiate between which servercommunicator is being used until we implement a different solution. I also implemented several portions of the serverservercommunicator's methods. none are tested yet, however.