Login and Registration feedback
Removed the System.out method calls to be replaced by a private helper method that creates a dialog if the input is unsuccessful.

-LoginScreen::presentErrorDialog:
this private helper method displays an error dialog to provide more precise information to the user on what went wrong when logging in or registering their account.

-LoginScreen::determineErrorMessage:
this private helper method returns an error message based on the success code provided.