Update array formatting
Update the formatting for the arrays so they are easier to read:
Each line is tabbed to make it easier to see that they are part of a larger statement, rather than a statement in and of themselves. 
The first line was brought down to match the rest so that the puzzle is easier to read. 
Comments have also been added to the lines to show what is causing `.isComplete()` to return `false`.