Fixed more Server/Client integration issues
Updated the habit screen to load habits and coin values from the server on launch if the server already has them for that user. This necessitated several new getters and setters and the reworking of the getCoins method in ServerServerCommunicator, so that the information would load in correctly. I also added a new listener in habit screen to allow for proper passing of the authentication token during the initialization.