Implemented Rudimentary TrueMock
In order to accommodate the testing sweet, I implemented an overloaded constructor that passes in a tcpAddress to connect to. I then created a TrueMock of the server that mimics the real server for the tests and sends back expected feedback. *WARNING* This code is a little unsafe, and I'm not sure it 100% works yet. Branch should not be merged to development.